Extract status explanation fields partial
